# bookaholic server/api/swagger.yaml
openapi: 3.0.2
info:
  title: Bookaholic API
  description: Bookaholic server offering access to books, reviews, users and related
    data
  termsOfService: http://addToShere
  contact:
    name: Carmine Alberto
    email: carmine.alberto@mail.polimi.it
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: localhost:8080
  description: API run offline - for testing purposes
- url: herokuURL
  description: Official API server where the application resides - for clients' consumption
...
paths:
...
  /orders:
    get:
      tags:
      - User
      summary: Get the user's orders
      operationId: getOrders
      parameters:
      - name: limit
        in: query
        description: The number of items to return.
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      - name: offset
        in: query
        description: Items to skip before starting to collect the response set.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          example: 5
          default: 0
      responses:
        200:
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/Order'
                x-content-type: application/json
        401:
          description: Missing JWT
      security:
      - jwtAuth: []
      x-swagger-router-controller: User
    post:
      tags:
      - User
      summary: Add a new order
      operationId: postToOrders
      responses:
        204:
          description: The request was carried out, no further information will be
            supplied
        401:
          description: Missing JWT
        403:
          description: 'Forbidden operation: no items in the cart'
      security:
      - jwtAuth: []
      x-swagger-router-controller: User
    x-swagger-router-controller: Order
  /profile:
    get:
      tags:
      - User
      summary: Get the user profile
      operationId: getProfile
      responses:
        200:
          description: Returns the authenticated user's data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Missing JWT
      security:
      - jwtAuth: []
      x-swagger-router-controller: User
    post:
      tags:
      - User
      summary: Create a new profile
      operationId: register
      requestBody:
        description: Data retrieved via a form to create a new user
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        201:
          description: Profile created successfully. User's data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        403:
          description: 'Forbidden action: the user already exists'
      x-swagger-router-controller: User
    x-swagger-router-controller: Profile
  /profile/{field}:
    put:
      tags:
      - User
      summary: Overwrite the specified resource
      operationId: updateField
      parameters:
      - name: field
        in: path
        description: Specify the field to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: string
          enum:
          - username
          - email
          - password
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        204:
          description: The request was carried out, no further information will be
            supplied
        400:
          description: Provided parameters do not match the requested format
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
        401:
          description: Missing JWT
      security:
      - jwtAuth: []
      x-swagger-router-controller: User
    x-swagger-router-controller: Profile
...
components:
  schemas:
...
    Order:
      required:
      - emission_time
      - order_id
      - status
      - total
      type: object
      properties:
        order_id:
          pattern: ^\d{10}$
          type: string
        emission_time:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - pending
          - dispatched
          - reservation
        total:
          minimum: 0.01
          type: number
          format: float
        details:
          minItems: 1
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Order_details'
      description: Object representing the an emitted order. We had thought about
        adding an additional attribute, address, and may do so in the future
      example:
        order_id: 1231314421
        emission_time: 2019-06-02T21:45Z
        status: reservation
        total: 47.9
        details:
        - {}
    User:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 3
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        email:
          type: string
          format: email
      description: Object representing a user. username is the identifier
      example:
        username: jace
        password: abcd
        email: jace.wayland@domain.com
...
    body:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        e-mail:
          type: string
          format: email
...
    Order_details:
      required:
      - amount
      - book_id
      - book_title
      - cover_type
      - price
      type: object
      properties:
        book_id:
          pattern: ^\d{10}$
          type: string
          format: book_id
        book_title:
          type: string
        cover_type:
          type: string
          enum:
          - hardcover
          - paperback
          - digital
        price:
          minimum: 0.01
          type: number
          format: float
        amount:
          minimum: 1
          type: number
          format: int32
...
  securitySchemes:
    jwtAuth:
      type: apiKey
      name: JWT
      in: cookie